chore: running all tests with address sanitizer and fixing issues refactoring: LocalVariableStorage - extracting repetitive code into inline function refactoring LVS: revised AddProperty() and DeleteProperty() methods refactoring: removed some superseded utilities testing: repaired some regions.txt files testing: extended ModelTopology_Test testing: fixed issues with destructor sequencing in PropertyHandle_Test testing: updated main() files for testing

+++ source/includes/model/LocalVariables.h
@@ -6,8 +6,8 @@ namespace csmp {
 /// Data class to store physical variable count at given placement
 struct LocalVariables {
 
-  // TODO: make all these uint16_t (the smallest size that can hold total datadepth in many-variable models)?
-  using int_type = uint32_t; ///< unsigned integer type that is big enough to hold 'totalDataDepth'
+  // TODO: investigate which type is best for performance
+  using int_type = uint16_t; ///< unsigned integer type that is big enough to hold 'totalDataDepth'
 
   LocalVariables()
     : scalars           (0),
